reconfigured packaging; including licenses

+++ compile_cli.m
@@ -16,51 +16,58 @@
 % You should have received a copy of the GNU Lesser General Public
 % License along with this library; if not, write to the Free Software
 % Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
-function compile_cli()
+function compile_cli(versno)
 % builds cli_plot_composite_maps.m  cli_plot_spectra.m  cli_solver.m
-    pkgdir='./bin';
-    builddir='./build';
-    if exist(pkgdir, 'dir')
-        rmdir(pkgdir,'s')
+    if (nargin<1)
+        versno='25a';
     end
-    if exist(builddir, 'dir')
-        rmdir(builddir,'s')
-    end
-    [~, ~, ~] = mkdir(pkgdir);
-    [~, ~, ~] = mkdir(builddir);
-    % cd build
-    Ext='';
-    Arch='';
+    exec_extension='';
+    architecture='';
     if ismac
-        Ext='.app';
+        exec_extension='.app';
         [~,m]=system('uname -m'); m=strtrim(m);
-        Arch=['mac_' m];
+        architecture=['mac_' m];
     elseif ispc
-        Arch='win'
-        Ext='.exe';
+        architecture='win'
+        exec_extension='.exe';
     else % assume linux
         [~,m]=system('uname -m'); m=strtrim(m);
-        Arch=['linux_' m];
+        architecture=['linux_' m];
     end
+    pkgname=sprintf('drsuite%s_%s',versno,architecture)
+    pkgdir=fullfile('.',pkgname);
+    if exist(pkgdir, 'dir'); rmdir(pkgdir,'s'); end
+    [~, ~, ~] = mkdir(pkgdir);
+    bindir=fullfile(pkgdir,'/bin')
+    if exist(bindir, 'dir'); rmdir(bindir,'s'); end
+    [~, ~, ~] = mkdir(bindir);
+    copyfile('LICENSE.txt',pkgdir)
+    copyfile('readme.md',pkgdir);
+    copyfile('ini2struct/license.txt',fullfile(pkgdir,'ini2struct_license.txt'))
+
+    builddir='./build';
+    if exist(builddir, 'dir'); rmdir(builddir,'s'); end
+    [~, ~, ~] = mkdir(builddir);
+
     srcfiles=[ "plot_avg_spectra", "plot_comp_maps",...
         "plot_spect_im", "plot_beta_sweep", ...
         "estimate_crlb","estimate_spectra", ...
           "create_phantom" ];
     for i=1:length(srcfiles)
         mfile=sprintf("%s.m",srcfiles(i));
-        efile=sprintf("%s%s",srcfiles(i),Ext);
+        efile=sprintf("%s%s",srcfiles(i),exec_extension);
         fprintf(1,"compiling %s to %s\n",mfile,efile);
         mcc("-m","-d",builddir,"-a","utilities",...
             "-a","solvers", "-a","ini2struct",...
             sprintf("%s.m",srcfiles(i)))
         if ~ispc
             sfile=sprintf("%s.sh",srcfiles(i));
-            dest=fullfile(pkgdir,sfile);
+            dest=fullfile(bindir,sfile);
             fprintf('updating %s...\n',sfile)
             writelines(change_runtime_version(fullfile('scripts',sfile)),dest);
             fileattrib(dest, '+x', 'all') 
         end
-        movefile(fullfile(builddir,efile),fullfile(pkgdir,efile))
+        movefile(fullfile(builddir,efile),fullfile(bindir,efile))
     end
-    zip(['drsuite_' Arch '.zip'],'bin')                
+    zip([pkgname '.zip'],pkgdir)                
 return
